follow redirects when prerendering
