+++ .changeset/angry-tigers-prove.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/kit': patch
+---
+
+Follow redirects when prerendering
