+++ packages/kit/src/core/adapt/prerender.js
@@ -1,10 +1,11 @@
 import { readFileSync, writeFileSync } from 'fs';
 import { dirname, join, resolve as resolve_path } from 'path';
-import { pathToFileURL, resolve, URL } from 'url';
+import { pathToFileURL, URL } from 'url';
 import { mkdirp } from '../../utils/filesystem.js';
 import { __fetch_polyfill } from '../../install-fetch.js';
 import { SVELTE_KIT } from '../constants.js';
 import { get_single_valued_header } from '../../utils/http.js';
+import { resolve } from '../../utils/url.js';
 
 /**
  * @typedef {import('types/config').PrerenderErrorHandler} PrerenderErrorHandler
@@ -191,6 +192,11 @@ export async function prerender({ cwd, out, log, config, build_data, fallback, a
 					log.warn(`${rendered.status} ${decoded_path} -> ${location}`);
 					writeFileSync(file, `<meta http-equiv="refresh" content="0;url=${encodeURI(location)}">`);
 					written_files.push(file);
+
+					const resolved = resolve(path, location);
+					if (resolved.startsWith('/')) {
+						await visit(resolved, path);
+					}
 				} else {
 					log.warn(`location header missing on redirect received from ${decoded_path}`);
 				}
