+++ packages/kit/src/runtime/server/page/load_node.js
@@ -1,6 +1,7 @@
 import { normalize } from '../../load.js';
 import { respond } from '../index.js';
 import { escape_json_string_in_html } from '../../../utils/escape.js';
+import { resolve } from '../../../utils/url.js';
 
 const s = JSON.stringify;
 
@@ -302,35 +303,4 @@ export async function load_node({
 		set_cookie_headers,
 		uses_credentials
 	};
-}
-
-const absolute = /^([a-z]+:)?\/?\//;
-
-/**
- * @param {string} base
- * @param {string} path
- */
-export function resolve(base, path) {
-	const base_match = absolute.exec(base);
-	const path_match = absolute.exec(path);
-
-	if (!base_match) {
-		throw new Error(`bad base path: "${base}"`);
-	}
-
-	const baseparts = path_match ? [] : base.slice(base_match[0].length).split('/');
-	const pathparts = path_match ? path.slice(path_match[0].length).split('/') : path.split('/');
-
-	baseparts.pop();
-
-	for (let i = 0; i < pathparts.length; i += 1) {
-		const part = pathparts[i];
-		if (part === '.') continue;
-		else if (part === '..') baseparts.pop();
-		else baseparts.push(part);
-	}
-
-	const prefix = (path_match && path_match[0]) || (base_match && base_match[0]) || '';
-
-	return `${prefix}${baseparts.join('/')}`;
 }
